move assertions code to the assertion executor component
